# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Code Hardener API
  version: 2.0.0
  description: Security assurance platform API with 70+ integrated scanners
servers:
  - url: http://localhost:7002/api/v1
    description: Local development
  - url: https://scan.codehardener.com/api/v1
    description: Production
paths:
...
  /scans:
    get:
      summary: List scans
      tags: [Scans]
      parameters:
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Scan list
    post:
      summary: Start a scan
      tags: [Scans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId: { type: string, format: uuid }
                profile: { type: string, enum: [quick, standard, comprehensive, security, api, performance, frontend, supply-chain, ai-security, ai-code-quality, database, chaos, pr, pre-commit, compliance, usability, unit-test, full] }
                scanners: { type: array, items: { type: string } }
              required: [projectId]
      responses:
        '201':
          description: Scan queued
  /scans/{id}:
    get:
      summary: Get scan details
      tags: [Scans]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Scan details with scanner results
...
  /auth/register:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
              required: [email, password, name]
      responses:
        '201':
          description: User registered
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        '200':
          description: Login successful with tokens
security:
  - devHeader: []
  - bearerAuth: []
components:
  securitySchemes:
    devHeader:
      type: apiKey
      in: header
      name: X-User-Id
      description: Dev mode auth bypass
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT